Added lots of Lt. Croy conversations
I went on the Starcruiser (2nd time) on March 30 to April 1.  I took lots of screenshots of my conversations, in particular with Lt. Croy, as I was following the first order path.  I've added as much as I could here!  

Note: Not sure how to test this - would love to have it tested before accepting to make sure there aren't any syntax errors that break the whole thing. :)  It should be right, but I could have made a mistake!